# internal/docs/swagger.yaml
basePath: /
definitions:
...
  models.BaseError:
    properties:
      error:
        example: something bad
        type: string
    type: object
...
  models.DeviceMetadata:
    properties:
      device_id:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
    type: object
  models.DeviceMetadataValue:
    properties:
      value:
        type: string
    type: object
...
info:
  contact:
    name: The Nexodus Authors
    url: https://github.com/nexodus-io/nexodus/issues
  description: This is the Nexodus API Server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Nexodus API
  version: "1.0"
paths:
...
  /api/devices/{id}/metadata:
    get:
      consumes:
      - application/json
      description: Lists metadata for a device
      operationId: GetDeviceMetadata
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceMetadata'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/models.BaseError'
      summary: Get Device Metadata
      tags:
      - Devices
    post:
      consumes:
      - application/json
      description: Delete all metadata for a device
      operationId: DeleteDeviceMetadata
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Metadata Key
        in: path
        name: key
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/models.BaseError'
      summary: Delete all Device metadata
      tags:
      - Devices
  /api/devices/{id}/metadata/{key}:
    get:
      consumes:
      - application/json
      description: Get metadata for a device
      operationId: GetDeviceMetadataKey
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Metadata Key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeviceMetadataValue'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/models.BaseError'
      summary: Get Device Metadata
      tags:
      - Devices
    post:
      consumes:
      - application/json
      description: Delete a metadata key for a device
      operationId: DeleteDeviceMetadataKey
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Metadata Key
        in: path
        name: key
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/models.BaseError'
      summary: Delete a Device metadata key
      tags:
      - Devices
...
securityDefinitions:
  OAuth2Implicit:
    authorizationUrl: https://auth.try.nexodus.127.0.0.1.nip.io/
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      user: ' Grants read and write access to resources owned by this user'
    type: oauth2
swagger: "2.0"